+++ src/main/java/com/projectmatching/app/domain/comment/dto/UserCommentDto.java
@@ -4,7 +4,9 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.projectmatching.app.domain.comment.entity.UserComment;
 import com.projectmatching.app.domain.liking.dto.UserCommentLikingDto;
+import com.projectmatching.app.domain.user.dto.UserInfo;
 import com.projectmatching.app.util.IdGenerator;
+import io.swagger.annotations.ApiModelProperty;
 import lombok.*;
 import org.springframework.beans.BeanUtils;
 
@@ -23,13 +25,22 @@ import static com.projectmatching.app.util.StreamUtil.mapToSet;
 public class UserCommentDto {
 
 
-    @JsonIgnore
-    private Long id = IdGenerator.number();
+    private final Long id = IdGenerator.number();
 
-    private String writer;
+
+    @ApiModelProperty(value = "댓글 작성자 정보")
+    private UserInfo userInfo;
+
+    @ApiModelProperty(value = "해당 댓글이 속한 유저 게시물의 id")
     private Long userId; //댓글이 속한 글의 id(유저)
+
+    @ApiModelProperty(value = "해당 댓글이 속한 댓글 id, 즉 부모 아이디가 없으면 일반 댓글, 있으면 대댓글")
     private Long parentId;
+
+    @ApiModelProperty(value = "댓글 비밀 여부")
     private Boolean secret;
+
+    @ApiModelProperty(value = "댓글 내용")
     private String content;
 
 
@@ -47,7 +58,7 @@ public class UserCommentDto {
         UserCommentDto userCommentDto = createEmpty();
         BeanUtils.copyProperties(userComment,userCommentDto);
         userCommentDto.userId = userComment.getUser().getId();
-        userCommentDto.writer = userComment.getUser().getName();
+        userCommentDto.userInfo = UserInfo.of(userComment.getUser());
         //부모가 있다면, 즉 대댓글이라면
         if(userComment.hasParent()){
             userCommentDto.parentId = userComment.getParent().getId();
